refactor(api): deixar o after seguro
- melhorar o after para que execute somente se for necessário evitando

+++ cypress/e2e/api/product/post.cy.js
@@ -48,7 +48,7 @@ describe('POST /produtos', () => {
   });
 
   after(() => {
-    cy.deleteProductAPI(token, productId);
-    cy.deleteUserAPI(userId);
+    if (productId) cy.deleteProductAPI(token, productId);
+    if (userId) cy.deleteUserAPI(userId);
   });
 });
